bff/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
strawberry-graphql==0.171.1
pulsar-client==3.1.0
aiopulsar-py==0.1.0a0
fastavro==1.7.3
requests==2.28.2
sse-starlette==1.3.3
pydantic==1.10.7